Extract inspection tab widget
